+++ packages/bridge-remote/src/RemoteBridgeAuthenticator.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Evolve\Bridge\Remote;
+
+use Evolve\Bridge\Contracts\BridgeError;
+use Psr\Http\Message\ServerRequestInterface;
+
+/**
+ * @experimental
+ */
+interface RemoteBridgeAuthenticator
+{
+    public function authenticate(ServerRequestInterface $request, RemoteBridgeInvocation $invocation): ?BridgeError;
+}