+++ tools/validate-prerelease-consumers.php
@@ -55,12 +55,14 @@ final class PrereleaseConsumerValidator
             $results[] = $this->runExpectedFailureCase($temp, $composer, 'Full-graph case E', $alphaRepositories, $lockedRuntimePackages, array(
                 'evolvephp/bridge-contracts' => '^2.0@alpha',
                 'evolvephp/bridge-psr' => '^2.0@alpha',
+                'evolvephp/bridge-remote' => '^2.0@alpha',
                 'evolvephp/dev-tools' => '^2.0@alpha',
                 'evolvephp/testing' => '^2.0@alpha',
             ));
             $results[] = $this->runExpectedSuccessCase($temp, $composer, 'Full-graph case F', $alphaRepositories, $lockedRuntimePackages, array(
                 'evolvephp/bridge-contracts' => '^2.0',
                 'evolvephp/bridge-psr' => '^2.0',
+                'evolvephp/bridge-remote' => '^2.0',
                 'evolvephp/dev-tools' => '^2.0',
                 'evolvephp/testing' => '^2.0',
             ), array('minimum-stability' => 'alpha', 'prefer-stable' => true), $this->expectedVersions($packages, self::ALPHA_VERSION));
@@ -68,6 +70,7 @@ final class PrereleaseConsumerValidator
             $results[] = $this->runExpectedSuccessCase($temp, $composer, 'Stable case H', $stableRepositories, $lockedRuntimePackages, array(
                 'evolvephp/bridge-contracts' => '^2.0',
                 'evolvephp/bridge-psr' => '^2.0',
+                'evolvephp/bridge-remote' => '^2.0',
                 'evolvephp/dev-tools' => '^2.0',
                 'evolvephp/testing' => '^2.0',
             ), array(), $this->expectedVersions($packages, self::STABLE_VERSION));